Move shared Supabase types into _shared/types and keep a thin supabase.ts barrel
